ダイアログが開いている時はこのコンポーネントのショートカットキーを無効化

+++ resources/js/Layouts/BaseBookMarkLayout.vue
@@ -171,13 +171,18 @@ export default {
         })
         //キーボード受付
         document.addEventListener('keydown', (event)=>{
-            // 削除ダイアログ呼び出し
-            if(this.$store.state.globalLoading === false){
+            // ダイアログが開いている時,読み込み中には呼ばせない
+            if( this.$store.state.globalLoading === false &&
+                this.$refs.deleteAlert.deleteDialogFlag === false &&
+                this.$refs.tagDialog.tagDialogFlag === false
+            ){
+                // 削除ダイアログ呼び出し
                 if (event.key === "Delete") {
                     this.$refs.deleteAlert.deleteDialogFlagSwitch()
                     return
                 }
 
+                // タグダイアログを開く
                 if ((event.ctrlKey || event.key === "Meta")
                 && event.altKey && event.code === "KeyT" ) {
                     event.preventDefault();
